backend/requirements.txt
aioredis==1.2.0
amqp==2.4.0
asn1crypto==0.24.0
async-timeout==3.0.1
attrs>=19.2.0
autobahn==19.6.2
Automat>=0.8.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
cffi==1.14.6
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==2.3.0
Django==4.0.1
django-cors-headers==3.11.0
django-debug-toolbar==1.11
djangorestframework==3.12.4
djangorestframework-simplejwt==3.3
drf-tracking==1.5.0
fabric==2.4.0
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
invoke==1.2.0
itypes==1.1.0
Jinja2==2.10
kombu==4.2.2.post1
MarkupSafe==1.1.0
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pytz==2018.9
redis==2.10.6
requests==2.21.0
# six==1.12.0
sqlparse==0.2.4
Twisted==21.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
